+++ src/app/auth/reset-password/reset-password-form.tsx
@@ -1,0 +1,42 @@
+'use client';
+
+import { Controller } from 'react-hook-form';
+
+import { Button } from '@/components/ui/button';
+import { Field, FieldGroup, FieldLabel, FieldError } from '@/components/ui/field';
+import { Input } from '@/components/ui/input';
+
+import useResetPasswordForm from './use-reset-password-form';
+
+export function ResetPasswordForm() {
+  const { form, onSubmit } = useResetPasswordForm();
+
+  return (
+    <form onSubmit={form.handleSubmit(onSubmit)} className="flex flex-col gap-6">
+      <FieldGroup>
+        <Controller
+          name="password"
+          control={form.control}
+          render={({ field, fieldState }) => (
+            <Field data-invalid={fieldState.invalid}>
+              <FieldLabel htmlFor="password">Password</FieldLabel>
+              <Input
+                {...field}
+                id="password"
+                aria-invalid={fieldState.invalid}
+                placeholder="••••••••"
+                autoComplete="off"
+              />
+              {fieldState.invalid && <FieldError errors={[fieldState.error]} />}
+            </Field>
+          )}
+        />
+        <Field>
+          <Button type="submit" className="cursor-pointer">
+            Save new password
+          </Button>
+        </Field>
+      </FieldGroup>
+    </form>
+  );
+}